Make layout content the point of forking
