+++ dotcom-rendering/src/layouts/FullPageInteractiveLayout.tsx
@@ -28,7 +28,7 @@ import type { ArticleDeprecated } from '../types/article';
 import type { Switches } from '../types/config';
 import type { FEElement } from '../types/content';
 import type { RenderingTarget } from '../types/renderingTarget';
-import { temporaryBodyCopyColourOverride } from './InteractiveLayoutDeprecated';
+import { temporaryBodyCopyColourOverride } from './interactives/InteractiveGridV1';
 import { interactiveGlobalStyles } from './lib/interactiveLegacyStyling';
 import { BannerWrapper, Stuck } from './lib/stickiness';
